# sstable: retain blockIter buffers across sync.Pool reuse