Retain the `blockIter.{fullKey,cached,cachedBuf}` (the dynamically
allocated buffers) across `sync.Pool` reuse of `singleLevelIterator` and
`twoLevelIterator`. This significantly reduces allocations which is the
point of pool allocation in the first place. This also obviates the need
for `blockIter.keyBuf` which has now been removed.

This provides a nice speedup on CRDB's short range (1-10 row)
`MVCCReverseScan` benchmarks:

```
name                                                         old time/op    new time/op    delta
MVCCReverseScan_Pebble/rows=1/versions=1/valueSize=8-16        3.17µs ± 3%    2.69µs ± 2%  -15.39%  (p=0.000 n=10+10)
MVCCReverseScan_Pebble/rows=1/versions=1/valueSize=64-16       5.07µs ± 4%    4.00µs ± 3%  -21.08%  (p=0.000 n=10+10)
MVCCReverseScan_Pebble/rows=1/versions=1/valueSize=512-16      4.58µs ± 2%    3.94µs ± 1%  -14.02%  (p=0.000 n=10+8)
MVCCReverseScan_Pebble/rows=10/versions=1/valueSize=8-16       5.56µs ± 5%    4.95µs ± 4%  -11.08%  (p=0.000 n=10+10)
MVCCReverseScan_Pebble/rows=10/versions=1/valueSize=64-16      7.84µs ± 4%    6.76µs ± 1%  -13.83%  (p=0.000 n=10+8)
MVCCReverseScan_Pebble/rows=10/versions=1/valueSize=512-16     9.31µs ± 2%    8.57µs ± 1%   -7.92%  (p=0.000 n=10+9)
MVCCReverseScan_Pebble/rows=100/versions=1/valueSize=8-16      22.9µs ± 3%    22.7µs ± 3%     ~     (p=0.075 n=10+10)
MVCCReverseScan_Pebble/rows=100/versions=1/valueSize=64-16     30.3µs ± 3%    29.6µs ± 4%   -2.13%  (p=0.035 n=9+10)
MVCCReverseScan_Pebble/rows=100/versions=1/valueSize=512-16    46.6µs ± 2%    46.0µs ± 2%     ~     (p=0.095 n=9+10)
```

Also verified that there was no impact on the forward scan benchmarks.